system/requirements.txt
keyboard
RPi.GPIO
paho-mqtt
pySerial
tinydb
tk_tools